Consider restarted pods as started and move to end of queue
Think on this a bit more... on one hand, a restarting pod can block
any other starting pods from releasing. On the other hand, the entire
queue of starting pods may restart for some reason and lead to node
thrashing.  It's probably more likely that only one pod will be
restarting, at least from our pods, but may be better to still
block the entire set of queued pods since there _should_ be
redundancy running elsewhere in the cluster.  However... this
means that the pods may be blocked indefinitely and require
sufficient alerting and manual intervention to push the pod
to the end of the queue again (where it can again start blocking
pods...).